Only initialise targets and device tree once
Commit 46aaa0b ("Moved pdbg target init/probe to later in software stack")
changed where the device tree and targets were initialised and probed so that we
could add a "--device" command to specify which device tree to use. This change
caused the probing to happen twice resulting in a segfault due to string
references changing.

This patch adds a static done flag to make sure we only init targets once. It's
probably not the most elegant solution but it will do for now as having the
`--device` is very useful for testing.

Signed-off-by: Alistair Popple <alistair@popple.id.au>